todo_backend/requirements.txt
django==5.2.4
djangorestframework==3.15.2
psycopg2-binary==2.9.9
redis==5.0.8
celery==5.4.0
django-celery-beat==2.8.1